% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummary.R
\name{createSummary}
\alias{createSummary}
\title{Get the summary details for the directory}
\usage{
createSummary(workFolder)
}
\arguments{
\item{workFolder}{The directory where the plpData and population are saved to}
}
\value{
Returns TRUE when finished and saves the summary csv  into the workFolder directory
names summary.csv
}
\description{
This function creates a csv summarising all the model results
}
\details{
Searching through the directory to extract the model evaluation for each model
}
