% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAllPredictionModels.R
\name{fitAllPredictionModels}
\alias{fitAllPredictionModels}
\title{Creates the models for each classifier for all outcomes}
\usage{
fitAllPredictionModels(workFolder, verbosity = INFO)
}
\arguments{
\item{workFolder}{The directory where the plpData and population are saved to}
}
\value{
Returns TRUE when finished and saves the models into the workFolder directory
in the subdirectory named models
}
\description{
This function creates the models for each classifier for all outcomes
}
\details{
The patient level prediction runPlp function is called for each outcome to train
each classifier
}
