% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGBMPredictionModels.R
\name{fitGBMPredictionModels}
\alias{fitGBMPredictionModels}
\title{Creates the gradient boosting machine models for all outcomes}
\usage{
fitGBMPredictionModels(workFolder)
}
\arguments{
\item{workFolder}{The directory where the plpData and population are saved to}
}
\value{
Returns TRUE when finished and saves the models  into the workFolder directory
in the subdirectory named gbmModels
}
\description{
This function creates the gradient boosting machine models for all outcomes
}
\details{
The patient level prediction runPlp function is called for each outcome to train
a gradient boosting machine model
}
