% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitKNNPredictionModels.R
\name{fitKNNPredictionModels}
\alias{fitKNNPredictionModels}
\title{Creates the K-nearest neighbours models for all outcomes}
\usage{
fitKNNPredictionModels(workFolder)
}
\arguments{
\item{workFolder}{The directory where the plpData and population are saved to}
}
\value{
Returns TRUE when finished and saves the models  into the workFolder directory
in the subdirectory named knnModels
}
\description{
This function creates the K-nearest neighbours models for all outcomes
}
\details{
The patient level prediction runPlp function is called for each outcome to train
a K-nearest neighbours model
}
