% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLassoPredictionModels.R
\name{fitLassoPredictionModels}
\alias{fitLassoPredictionModels}
\title{Creates the lasso regularised logistic models for all outcomes}
\usage{
fitLassoPredictionModels(workFolder)
}
\arguments{
\item{workFolder}{The directory where the plpData and population are saved to}
}
\value{
Returns TRUE when finished and saves the models  into the workFolder directory
in the subdirectory named lrModels
}
\description{
This function creates the lasso regularised logistic models for all outcomes
}
\details{
The patient level prediction runPlp function is called for each outcome to train
a lasso regularised logistic regression model
}
