% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNaiveBayesPredictionModels.R
\name{fitNaiveBayesPredictionModels}
\alias{fitNaiveBayesPredictionModels}
\title{Creates the naive bayes models for all outcomes}
\usage{
fitNaiveBayesPredictionModels(workFolder)
}
\arguments{
\item{workFolder}{The directory where the plpData and population are saved to}
}
\value{
Returns TRUE when finished and saves the models  into the workFolder directory
in the subdirectory named nbModels
}
\description{
This function creates the naive bayes models for all outcomes
}
\details{
The patient level prediction runPlp function is called for each outcome to train
a naive bayes model
}
