% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitResults.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(workFolder, dbName, N = NULL, localFolder = NULL)
}
\arguments{
\item{workFolder}{Name of local folder to place results}

\item{dbName}{Database name used in the zipName}

\item{N}{The minimum count size for summary data to be exported for data sensitivity}

\item{localFolder}{Name of the folder where sensitive data gets moved}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
